requirements.txt
faiss-cpu
transformers
sentence-transformers
torch
numpy
beautifulsoup4
lxml
html2text
requests
tqdm
fastapi
uvicorn
python-multipart
groq
pydantic
rich
typing-extensions
